fix: compatibility and return with error
The program wouldn't work on some PCs: the SDL_Renderer wouldn't be generated with the SDL_RENDERER_ACCELERATED flag on some PCs, so if that doesn't work, I try to generate it with the SDL_RENDERER_SOFTWARE flag. If it still doesn't work, then I return NULL (init_gui()). Also I made the program return 1 if init_gui() fails